Add random data output to decode success tests
