+++ library/encoding-base32/src/commonTest/kotlin/io/matthewnelson/encoding/base32/Base32DefaultUnitTest.kt
@@ -40,6 +40,23 @@ class Base32DefaultUnitTest: BaseNEncodingTest() {
 
     override val decodeSuccessDataSet: Set<Data<String, ByteArray>> = setOf(
         decodeSuccessHelloWorld,
+        Data(
+            raw = "KNQWY5DFMRPV76DEF2232QQS4NFDHX7K33NZK5K3K7AWZVWK7ZU5DAO32PEYH5ERWNGCOHXOO5LBGXJYMBZAOHSHJ4AP5H6NWGRVGZI=",
+            expected = ("53 61 6c 74 65 64 5f 5f f8 64 2e b5 bd 42 12 e3 4a 33 df ea de db 95 75 5b 57 c1 " +
+                    "6c d6 ca fe 69 d1 81 db d3 c9 83 f4 91 b3 4c 27 1e ee 77 56 13 5d 38 60 72 07 1e 47 " +
+                    "4f 00 fe 9f cd b1 a3 53 65").decodeHexToByteArray()
+        ),
+        Data(
+            raw = "KNQWY5DFMRPV7RXZCOJOT3QZ3KFMCGGJNSPKK3O3V53JPXIV2LNFYV35OUOF42NPUWHGLQGHH2QIO===",
+            expected = ("53 61 6c 74 65 64 5f 5f c6 f9 13 92 e9 ee 19 da 8a c1 18 c9 6c 9e a5 " +
+                    "6d db af 76 97 dd 15 d2 da 5c 57 7d 75 1c 5e 69 af a5 8e 65 c0 c7 3e a0 " +
+                    "87").decodeHexToByteArray()
+        ),
+        Data(
+            raw = "KNQWY5DFMRPV744USBXPVHIQ6GY57YAPIAKQCOUX5QPUROTWVWUQ====",
+            expected = ("53 61 6c 74 65 64 5f 5f f3 94 90 6e fa 9d 10 f1 b1 df e0 0f 40 15 01 " +
+                    "3a 97 ec 1f 48 ba 76 ad a9").decodeHexToByteArray()
+        ),
         Data(raw = "========", expected = ByteArray(0), message = "Decoding a String containing only padding '=' should return an empty ByteArray"),
         Data(raw = "MY======", expected = "f".encodeToByteArray()),
         Data(raw = "MY", expected = "f".encodeToByteArray(), message = "Stripped padding should decode"),
